реализовал и протестировал основные классы

+++ 05_command/DocumentEditor/Image.cpp
@@ -6,13 +6,15 @@ using namespace std;
 using namespace boost;
 using namespace filesystem;
 
-CImage::CImage(const string & filePath, int width, int height, CHistory & history)
-	:m_fileName(path(filePath).filename().generic_string()),
-	m_mustDelete(false),
+CImage::CImage(const string & filePath, int width, int height)
+	:m_mustDelete(false),
 	m_width(width),
-	m_height(height),
-	m_history(history)
+	m_height(height)
 {
+	if (!exists(path(filePath)))
+		throw invalid_argument("Invalid file path or not accessible file");
+
+	m_fileName = path(filePath).filename().generic_string();
 	create_directory(path("images"));
 	copy_file(path(filePath), path("images") / m_fileName, copy_option::overwrite_if_exists);
 }
@@ -40,8 +42,8 @@ int CImage::GetHeight() const
 
 void CImage::Resize(int width, int height)
 {
-	m_history.AddAndExecuteCommand(
-		make_unique<CResizeImageCommand>(m_width, m_height, width, height));
+	m_width = width;
+	m_height = height;
 }
 
 void CImage::MustDelete(bool mustDelete)
